2/3 functional
Got the email check done. The system now detects newly breached emails and adds them to the database.
Checks all breached emails for new breaches, adding them to a list if they need updating.

I also re-formatted the API requesting so that we can better account for unhappy response codes, and to add in recursion for when the API needs a cool down time.