Development (#1)
* Working New Identity assessment container

* Before we break things

This a a checkpoint before we break things

* API check on startup

Made it so that the API gets called on startup to ensure the connection is viable.

* Created the Skeleton of the POC

Created the skeleton of the POC. Adding in the initial elements for API and database fetching.

I also left a couple of pseudo code block so Sahim can see where things are planned to go

* 2/3 functional

Got the email check done. The system now detects newly breached emails and adds them to the database.
Checks all breached emails for new breaches, adding them to a list if they need updating.

I also re-formatted the API requesting so that we can better account for unhappy response codes, and to add in recursion for when the API needs a cool down time.

* Finished the email update

Finished up the email update section.
Now all that's left is to generate the report and finalise the POC

* Finished the POC

Finally. After a long day of non-stop programming it is complete.
The system can fully check the domain with HIBP. detect the differences with the local database, correctly make updates, and report all the changes in a text file.

* New Breaches Functionality

Added in the functionality for checking with the HIBP API for new breaches. This ensures that we aren't doing domains scans when there's no new information available.

* Retry-After header

Changed the way time_to_sleep was collected so that different messages with code:429 didn't crash the system.

* Forgot to add in domain functionality

I forgot to add in the thing that extracted the dict of breached emails, when receiving a successful API return.
Oops

* Docker push

* removed test